add restaurant authorization service to check privilages to delete and update restaurants and dishes

+++ Restaurants.Application/Dishes/Commands/DeleteDishesForRestaurant/DeleteDishesForRestaurantCommandHandler.cs
@@ -1,7 +1,9 @@
 using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Constans;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Dishes.Commands.DeleteDishesForRestaurant;
@@ -9,10 +11,14 @@ namespace Restaurants.Application.Dishes.Commands.DeleteDishesForRestaurant;
 public class DeleteDishesForRestaurantCommandHandler(
     ILogger<DeleteDishesForRestaurantCommandHandler> logger,
     IRestaurantsRepository _restaurantsRepository,
-    IDishesRepository _dishesRepository
+    IDishesRepository _dishesRepository,
+    IRestaurantAuthorizationService restaurantAuthorizationService
 ) : IRequestHandler<DeleteDishesForRestaurantCommand>
 {
-    public async Task Handle(DeleteDishesForRestaurantCommand request, CancellationToken cancellationToken)
+    public async Task Handle(
+        DeleteDishesForRestaurantCommand request,
+        CancellationToken cancellationToken
+    )
     {
         logger.LogWarning(
             "Delete dishes for repository with id {RestaurantId}",
@@ -20,15 +26,21 @@ public class DeleteDishesForRestaurantCommandHandler(
         );
 
         var restaurant = await _restaurantsRepository.GetByIdAsync(request.RestaurantId);
-        if(restaurant==null){
-            throw new NotFoundException(nameof(Restaurant),request.RestaurantId.ToString());
+        if (restaurant == null)
+        {
+            throw new NotFoundException(nameof(Restaurant), request.RestaurantId.ToString());
         }
         // it generate separate delete sql query for each dish
         // restaurant.Dishes.Clear();
         // await _restaurantsRepository.SaveChangesAsync();
 
+        if (!restaurantAuthorizationService.Authorize(restaurant, ResourceOperations.Delete))
+        {
+            throw new ForbiddenException(
+                $"User not authorized to delete dish from restaurant with {restaurant.Id} id"
+            );
+        }
         //below method behave the same :: generate separate delete sql query for each dish
         await _dishesRepository.Delete(restaurant.Dishes);
-
     }
 }
